pyproject.toml
[project]
name = "usdaf-trove-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hikari>=2.3.3",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "web3>=7.12.0",
]
